requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-allauth==0.44.0
djangorestframework==3.12.2
django-cors-headers==3.7.0